Put the main modules in their places and set up connection
The main modules are in place.
The links in .sln and .csproj have been changed to correct ones.
The project is not launching yet.
Communication with Docker is not configured.
The schemas in the database are not separated.
The tests cause errors and it has not yet been decided where to store them.